requirements-dev.txt
-r requirements.txt
httpx
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
pytest-mock==3.14.0
mongomock-motor==0.0.34
pylint